requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
drf-spectacular==0.27.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0